pyproject.toml
[tool.poetry]
name = "sphiwe"
version = "0.1.0"
description = ""
authors = ["Nazire Mathe <nazireemathe@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
django = "4.2.11"
django-split-settings = "^1.3.1"
dotmap = "^1.3.30"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
